git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@3271 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/SVertexManipulator.h
@@ -67,7 +67,7 @@ namespace scene
 			video::SColor high) : Threshold(threshold), Low(low), High(high) {}
 		void operator()(video::S3DVertex& vertex) const
 		{
-			vertex.Color = (vertex.Color.getAverage()>Threshold)?High:Low;
+			vertex.Color = ((u8)vertex.Color.getAverage()>Threshold)?High:Low;
 		}
 	private:
 		u8 Threshold;
